More sonar cleanups
