+++ src/main/java/org/stt/gui/jfx/Tooltips.java
@@ -3,11 +3,11 @@ package org.stt.gui.jfx;
 import javafx.scene.Node;
 import javafx.scene.control.Tooltip;
 
-public abstract class Tooltips {
+class Tooltips {
     private Tooltips() {
     }
 
-    public static void install(Node node, String tooltip) {
+    static void install(Node node, String tooltip) {
         Tooltip.install(node, new Tooltip(tooltip));
     }
 }
